Moved source code to src directory. Fixed path issues
